+++ src/features/notes/notesSlice.js
@@ -3,8 +3,8 @@ import { createAsyncThunk } from "@reduxjs/toolkit";
 import { baseUrl } from "../../app/shared/baseUrl";
 
 export const fetchNotes = createAsyncThunk("notes/fetchNotes", async () => {
-  // item is in double quotes when retrieved ""item""
-  const user = localStorage.getItem("user").replace(/['"]+/g, "");
+  const user = JSON.parse(localStorage.getItem("user"));
+
   if (user === "offline") {
     return JSON.parse(localStorage.getItem("notes") ?? "[]");
   } else {
@@ -28,8 +28,36 @@ export const postNote = createAsyncThunk(
       return Promise.reject("empty input");
     }
 
-    const response = await fetch(baseUrl + "notes", {
-      method: "POST",
+    const user = JSON.parse(localStorage.getItem("user"));
+
+    if (user === "offline") {
+      dispatch(addNote(note));
+    } else {
+      const response = await fetch(baseUrl + "notes", {
+        method: "POST",
+        body: JSON.stringify(note),
+        headers: { "Content-Type": "application/json" },
+        credentials: "include",
+      });
+
+      if (!response.ok) {
+        return Promise.reject(response.status);
+      }
+
+      const data = await response.json();
+      dispatch(addNote(data));
+    }
+  },
+);
+
+export const updateNote = createAsyncThunk("notes/updateNote", async (note) => {
+  const user = JSON.parse(localStorage.getItem("user"));
+
+  if (user === "offline") {
+    return note;
+  } else {
+    const response = await fetch(baseUrl + `notes/${note._id}`, {
+      method: "PUT",
       body: JSON.stringify(note),
       headers: { "Content-Type": "application/json" },
       credentials: "include",
@@ -38,37 +66,27 @@ export const postNote = createAsyncThunk(
     if (!response.ok) {
       return Promise.reject(response.status);
     }
-
-    const data = await response.json();
-    dispatch(addNote(data));
-  },
-);
-
-export const updateNote = createAsyncThunk("notes/updateNote", async (note) => {
-  const response = await fetch(baseUrl + `notes/${note._id}`, {
-    method: "PUT",
-    body: JSON.stringify(note),
-    headers: { "Content-Type": "application/json" },
-    credentials: "include",
-  });
-
-  if (!response.ok) {
-    return Promise.reject(response.status);
   }
 });
 
 export const deleteNote = createAsyncThunk("notes/deleteNote", async (note) => {
-  const response = await fetch(baseUrl + `notes/${note._id}`, {
-    method: "DELETE",
-    credentials: "include",
-  });
+  const user = JSON.parse(localStorage.getItem("user"));
 
-  if (!response.ok) {
-    return Promise.reject(response.status);
-  }
+  if (user === "offline") {
+    return note;
+  } else {
+    const response = await fetch(baseUrl + `notes/${note._id}`, {
+      method: "DELETE",
+      credentials: "include",
+    });
 
-  if (response.ok) {
-    return { _id: note._id };
+    if (!response.ok) {
+      return Promise.reject(response.status);
+    }
+
+    if (response.ok) {
+      return { _id: note._id };
+    }
   }
 });
 
@@ -83,10 +101,24 @@ const notesSlice = createSlice({
   initialState,
   reducers: {
     addNote: (state, action) => {
-      const newNote = {
-        ...action.payload,
-      };
-      state.notesArray.push(newNote);
+      const user = JSON.parse(localStorage.getItem("user"));
+
+      if (user === "offline") {
+        const newNote = {
+          _id: state.notesArray.length + 1,
+          ...action.payload,
+        };
+
+        state.notesArray.push(newNote);
+        let localArr = JSON.stringify(state.notesArray);
+
+        localStorage.setItem("notes", localArr);
+      } else {
+        const newNote = {
+          ...action.payload,
+        };
+        state.notesArray.push(newNote);
+      }
     },
   },
   extraReducers: {
@@ -111,13 +143,43 @@ const notesSlice = createSlice({
       );
     },
     [deleteNote.fulfilled]: (state, action) => {
-      const deletedNoteId = action.payload._id;
+      const user = JSON.parse(localStorage.getItem("user"));
 
-      const updatedNotesArray = state.notesArray.filter(
-        (note) => note._id !== deletedNoteId,
-      );
+      if (user === "offline") {
+        let notes = JSON.parse(localStorage.getItem("notes"));
 
-      state.notesArray = updatedNotesArray;
+        let filtered = notes.filter((note) => note._id !== action.payload._id);
+        state.notesArray = filtered;
+
+        let stringArr = JSON.stringify(filtered);
+        localStorage.setItem("notes", stringArr);
+      } else {
+        const deletedNoteId = action.payload._id;
+
+        const updatedNotesArray = state.notesArray.filter(
+          (note) => note._id !== deletedNoteId,
+        );
+
+        state.notesArray = updatedNotesArray;
+      }
+    },
+    [updateNote.fulfilled]: (state, action) => {
+      const user = JSON.parse(localStorage.getItem("user"));
+
+      if (user === "offline") {
+        let notes = JSON.parse(localStorage.getItem("notes"));
+
+        const newNotes = notes.map((note) => {
+          if (note._id === action.payload._id) {
+            return { ...note, text: action.payload.text };
+          }
+          return note;
+        });
+
+        let result = JSON.stringify(newNotes);
+        localStorage.setItem("notes", result);
+        state.notesArray = newNotes;
+      }
     },
   },
 });
